refactor: Remove duplicate type declarations and add tests for EmailMarketerController and EmailMarketerService
