+++ src/email-marketer/test/email-marketer.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { EmailMarketerController } from './email-marketer.controller';
+import { EmailMarketerController } from '../email-marketer.controller';
 
 describe('EmailMarketerController', () => {
   let controller: EmailMarketerController;
