+++ src/utils/inferredTypes.ts
@@ -1,0 +1,13 @@
+export interface EmailGenerationResponse {
+  event_name: 'email_generated';
+  message: string;
+  status: 'success';
+  username: 'mastraAiemailgen';
+}
+
+export type EmailGeneration = {
+  event_name: 'email_generated';
+  message: string;
+  status: 'success';
+  username: 'mastraAiemailgen';
+};
